// .worktrees/contract-analytics-filters/src/types.ts
export type ApiError = {
  status: boolean;
  message: string;
};

export type ApiData<T = unknown> = {
  status: boolean;
  message: string;
  data: T;
};

export type ApiList<T = unknown> = {
  total: number;
  page: number;
  limit: number;
  data: T[];
};

export type ApiResponse<T = unknown> = AxiosResponse<ApiData<T>>;
export type ApiResponseError = AxiosError<ApiError>;